Add files via upload
The search parameter is entered from the command line. Has to be one string, so enter as so: "Val+to+search". Script will parse the string.

usage: python automated_googleSearch.py Search+Bob+and+Alice

Also, the Chrome driver has to be installed in order to run the script properly with Selenium. You need the right exe file that matches the Chrome version on your computer